pkg/generator: Give docker access to GOPATH if SELinux is enforcing
The code generator container was attempting to write into my GOPATH which is
labeled as part of my home directory, and thus docker containers are not, by
default, given access. This patch mountes the volume with `:Z` which will
cause docker to relabel the volume granting access to the docker container.

Fixes #213