test/e2e: general cleanup and renaming of variables
This renames many variables in the e2e test to make the naming more
consistent and thus make the code more readable. It also makes a
small change to writing the new cr.yaml file to make it simpler and
changes the delete options for the cr's deletion to make them the
same as the rest of the deletion process.