test/e2e: test `up local` command in e2e test
Previous versions of this test only tested the operator running
in the cluster, not when running local by using the `up local`
subcommand. This commit makes both methods be tested in parallel. To
determine unique namespaces and keep track of cleanup functions,
the context.go file from the prometheus operator was added to
the sdk e2e test framework in the commit as well.