test/e2e: move all resource creation to resource_creator
This moves all the resource creation (which is currently the
CreateFromYAML and namespace functions) into resource_creator.go.
This allows these functions to be called under TestCtx, which now
stores the namespace and a CRClient that can be retrieved by these
functions when needed instead of being regenerated. The functions
also now have access the Global framework variable and thus don't
need to have a kubeclient and kubeconfig be passed as arguments
when called.