sdk: add support for per-Informer Handlers
Today using the SDK it is not possible to set up multiple handlers (e.g. a handler per resource type) in the same binary because all Informers use the global Handler set by `sdk.Handle`. With this change, a user can create an Informer that will invoke a specific handler. We also add a sdk.WatchWith method which mirrors sdk.Watch but allows the user to provide a Handler for the underlying Informer.