pkg/generator: Add prune to Gopkg.toml template (#413)
The code that gets vendored by dep by default is huge. We can
leverage the functionality given by dep viz. prune. This will remove
all the non-go code and unused packages.

It makes an exception to the k8s.io/code-generator repository. Since
we use the scripts from this repository to generate code.

Ref: https://golang.github.io/dep/docs/Gopkg.toml.html#prune