test/e2e/memcached_test.go: fix travis on master
On master, TRAVIS_PULL_REQUEST_SLUG is still set, but to "". This
breaks dep if it is run after a dependency change occurs, which
happens once we add the test directory. To fix this, we check
if TRAVIS_PULL_REQUEST_SLUG is "" and we also only run `dep ensure`
once after all the files are up to date instead of once before
and once after.