pkg/test: create new frameworkClient
This commit create a new FrameworkClient type accesible as
Global.Client. It is a simple wrapper of the controller-runtime's
client with the exception of Create, which now adds cleanup
functions. This also changes "Finalizer" to "Cleanup" and trims
the Framework struct to only include variables that are useful
to the user. All other variables required in the framework are
now separate, unexported variables.