pkg/test: add new CreateWithFinalizer helper function
This commit adds a new helper function that creates a provided
object and then creates a finalizer for the function to make sure
it is cleaned up properly. This is especially useful for single
namespace testing. The finalizer has a wait.PollImmediate to make
sure that the resource is fully deleted before returning