test/e2e/memcached_test.go: fix e2e testing on local machines (#548)
PR #525 changed the way that dep worked to fix dependency issues
when doing e2e tests on travis. The PR breaks e2e testing on local
machines though, as the tests will just use the master branch
instead of a local branch. This commit symlinks the local sdk
into vendor on non-travis tests. The tests will still fail on local
machines if there is a dependency change, but we are unable to fix
this on local machines unless `dep` adds local repo support.