commands/{build,cmdutil},pkg/generator: remove build.sh
This commit moves all functionality of build.sh into the build.go
file to embed the functionality into the binary. It also allows the
binary to generate the files needed when making tests (Dockerfile
and go-test.sh) in case someone tries to build testing enabled
images using a project made before that functionality was added.

See #541