moves controller creation and startup out of ansible's main.go
main.go should be as minimal as is reasonable, so this moves startup logic to a
more appropriate place. It also slightly changes the proxy startup code to
better match controller startup.

The new error handling avoids a race between the proxy being available and
ansible trying to use it.