proposal:  User-driven Ansible Operator status management (#687)
Signed-off-by: Fabian von Feilitzsch <fabian@fabianism.us>

**Description of the change:**
Add a proposal to allow the users to manage the status of the CR from Ansible when using Ansible Operator

**Motivation for the change:**
Users have requested the ability to have more fine-grained control of the content of the status. This proposal is intended to begin a discussion on the proper way to surface that functionality.