pkg/scaffold/cmd.go: Skip erroring out on Service
If the Service could not be created we just log the error instead of
erroring out and exiting. This error could occur when there is more than
one pods such as with leader election. In that case the new pod will
expose the metrics via the already created Service object.