pkg/(ansible|helm)/flags,commands,test: centralizing zap flagset setup
Moves the zap flagset initialization into the ansible and helm flag
packages so that they are reused in all ansible and helm operator
scenarios (e.g. `up local`, `run`, and the helm binary)

Also reverts parts a previous commit (975eddc) that added flags to the
ansible and helm operators imported by other libraries because it caused
the entire set of "testing" package flags to be added (from
`operator-sdk` dependencies).