pkg/helm: new watch file option to support watching dependent resources
This commit adds a new option to the watches.yaml file called
`watchDependentResources`, which enables users to configure whether
resources defined by Helm releases are watched for reconciliation.

This commit also decouples the loading of the watch file from the
creation of manager factories by refactoring the loading of the watch
file into a separate `watches` package. This makes it easier to add new
watch configurations that are outside the scope of the release manager
factory.

Lastly, this commit removes helpers that allowed watches to be
configured from environment variables. Now that multiple watches are
supported and new watch-specific configurations are being added, loading
from the environment does not provide the necessary flexibility.