Expose controller-runtime metrics (#786)
* pkg/scaffold/cmd.go: Skip erroring out on Service

If the Service could not be created we just log the error instead of
erroring out and exiting. This error could occur when there is more than
one pods such as with leader election. In that case the new pod will
expose the metrics via the already created Service object.

* pkg/*: Create new client for Service get/create

Due to the cache not being started at the time when we attempt to query
for the Service, we instead create a new client in a similar way as we
do with leader elections.

* CHANGELOG: Document metrics addition