hack/ci: Add check for valid commit range (#919)
**Description of the change:**
In cases when PRs were force pushed and the commit range was not
valid, the PR job would still pass as we did catch the error. This adds
a check and exits in case of an invalid commit range.

**Motivation for the change:**
Ran into a weird error when running tests, where the commit didn't exist anymore(I think), as it exited the `check-doc-only-update.sh` with `true` which in turn was passing my tests.