test/test-framework/deploy/operator.yaml: update image (#1017)
The sdk's subcommand tests rely on a very old memcached example that
has not been updated in 4 months. I've pushed up a new memcached
operator image based on sdk v0.4.0. The old operator image did not
support the status resource, so the new operator and manifests have
been updated to work with subresources.