hack,pkg/ansible/proxy: move ansible proxy test to e2e
The ansible proxy test in pkg/ansible/proxy requires a running
cluster, and thus should not be run as a unit test. This commit
marks the test to be skipped when go test is run in short mode
and will instead be run as part of test/e2e/ansible. This commit
also disables go test caching