pkg,test: Add deployments/finalizers to roles
Because the Service now adds an OwnerReference wherever
OwnerReferencesPermissionEnforcement is enabled this role is needed
to be able to create an OwnerReference in the Service that points to the
Deployment.

Because we determine the owner of the Service dynamically, this role
would need to be adjusted by the user to fit whatever type of deployment
resource they use.