pkg/test: remove dependence on ctx.t
The test library has functions that are useful outside of `go test`,
such as `createFromYAML`, `ctx` resource deletion, and auto setting
of resource cleanup when using the framework's client. The only
function that required `t` and would fail if it did not exist was
`Cleanup`, as other function could handle if it was nil. This
commit removes `Cleanup`'s dependence on `t`