log: allow different log levels (#1210)
* log: allow different log levels

With this feature we can use --zap-level 10 and this will allow
log.V(10).Info("hello world") to be seen in the logs. Prior to this
change debug (log.V(1)) was the lowest value attainable.

We also disable sampling when using an integer less than -1
(i.e. beyond debug). The zap sampling code assumes the lowest number for
sampling is DebugLevel (-1) which isn't the case in this scenario.
Disabling sampling makes sense at this point because sampling is used
for better performing logs which isn't required for trace level type
logs.

We also handle negative values being passed in.

* log: add unit test for flags. Fix error value.
* support more levels & remove debug level warning
* add logger test
* do not allow negative log levels at the CLI
* disable sampling later in the call based on review
* add note about disabling sampling to help
* update sampling doc, fix log level help
* make sample text consistent with log level