docs: Make first ref to quay.io link to https://quay.io (#1220)
* docs: Make first ref to quay.io link to https://quay.io

In the off chance that someone hasn't has only heard of a single
container registry (eg Docker Hub), provide a hyperlink to quay.io on
its first mention. Plus, having the link provides for a little less
typing for people to get to the site.

* docs: Adjust quay.io link style

Addresses PR feedback (#1220) to change link style to
`[anchor text][link ref]` with a `[link ref]: https://quay.io` at the end of
the document.