pkg/helm,hack/tests: slim down release information in Helm CR status
With the Helm storage backend moved to in-cluster secrets, it is no
longer necessary to store the full release state on the CR status.
This commit removes the full release status from the status condition
and adds a new "deployedRelease" field on the top level status that
contains just the release name and manifest.