faq + explanation for `The resourceVersion for the provided watch is too old` warning
It would be helpful to have a variety of commonly asked operator-sdk questions, starting with `The resourceVersion for the provided watch is too old`. This question appears in a variety of blogs and Github issues but is not posted anywhere in our docs - many people think this is a bug.