Fix log message so it doesn't cause an error
Without this change, I get an error about "odd number of arguments
passed as key-value pairs for logging" in the case where this code is
invoked (I hit it when running the operator with `operator-sdk up local`).

I only hit it in the cmd.go for my Go operator, but I presume the same
thing is needed for the cmd_test.go and helm/run.go.