Add openshift subcommand to generate
This currently generates openshift specific manifest files, but the
command can be reused later on for any openshift specif generation.

The manifest that are generated only work in openshift clusters as they
have specific annotations to generate TLS certs and secure the metrics
endpoints that are exposed by the operator.