pkg/helm: deprecate UID-based release name, use CR name instead
This changes the release name for new CRs to be the CR name. Releases for
existing CRs will continue to use the legacy name. If a CR is created, and a
release already exists with the same name from a CR of a different kind, an
error is returned, indicating a duplicate name.

The reason for this change is based on an interaction between the Kubernetes
constraint that limits label values to 63 characters and the Helm convention
of including the release name as a label on release resources.

Since the legacy release name includes a 25-character value based on the
parent CR's UID, it leaves little extra space for the CR name and any other
identifying names or characters added by templates.