Allow additional labels on ServiceMonitor creation
When provisioning a Prometheus server with a Prometheus CR, one often
specifies a ServiceMonitorSelector.In this case, this Prometheus will only
scrape metrics from endpoints that are specified in ServiceMonitors
that are selected based on the labels specified in the
ServiceMonitorSelector.

Without the ability to set a label on the ServiceMonitor, then many
people would have to either not use the utility function from here, or
update the resource after creation.