metric: add an updater function to CreateServiceMonitors
With this addition it would be possible to provide a list of updater
functions to modify the ServiceMonitor after they are created and
before they are added to k8. This can be used for scenarios like
updating the labels on the ServiceMonitor, changing metadata or updating
exposed endpoints (service port); or literally any other change that a
user might need.

Signed-off-by: Alex Lourie <djay.il@gmail.com>