Allow up local detection by operator
When developing locally, it would be convient to detect this in
the actual operator code so things like metrics can be bypassed.

With the OPERATOR_UP_LOCAL=true, operators can skip certain
bootstrapping tasks like attempting to register with a metrics service
while still maintaining logic for registering failures in production.