hack/ci/marker: update marker
Update marker to v0.8.0. As before, this was built in a CentOS 7
VM, which allows it to run in both Travis and API-CI, and then
was stripped, then compressed by UPX.

Marker v0.8.0 fixes the `_` issue that was previously causing
problems in some PRs that changed markdown files, adds support
for Github style checklists using `[]`, and adds a --exclude flag.

The other changes in the PR are changing the Makefile to run marker
on the entire project tree and fixes a couple of new broken
references/URLs found by marker.