pkg/test/e2eutil: wait for a deployment to have >= n replicas (#2248)
Before waitForDeployment was waiting until the exact number of replicas
were available. Now, wait for at least that number of replicas to be
available.

scenario:
I'm using an operator with 3 replicas, but only 1 is needed for tests to
work. If all 3 replicas were available before waitForDeployment started
then it would wait until it timed out because 3 does not equal 1. Now,
since 3 replicas is greater than the 1 needed for tests waitForDeployment
successfully waits.