pkg/test/client: retry cleanup function if cluster is temporarily unavailable
There are instances where the cluster could be temporarily unavailable.
e.g. when etcd is doing leader re-election.

    https://search.svc.ci.openshift.org/?search=etcdserver%3A+leader+changed&maxAge=336h&context=2&type=all

While this is not a very normal scenario, it would be good to be lenient
on the cleanup side of things and retry if such a case happens.

This will be reflected as Timeout or Unavailable errors coming from etcd.
This patch proposes to retry with backoff in an error happens. After a
set number of retries, it'll return the appropriate error.