pkg/test/context.go: create an id that is a valid namespace name
Before this change it was possible for `TestCtx.GetNamespace()` to create a
namespace with too long of a name, thus getting an error from Kubernetes
saying the namespace name exceeded 63 characters.

Now set the id to "osdk-e2e-UUID", so that `TestCtx.GetNamespace()` is
guaranteed to have a namespace name be 63 characters or less, which is a
valid namespace name.

Fixes #2101