cli: Allow build to skip building the image
Allow the user to --skip-image, in which case they will not be
required to give an image name on the command line

This change allows the user to build the go and the dockerfile but not run
any command to build the image. They can use the executable or an
unsupported image builder to build the image.

Fixes #2568