Add --skip-cleanup-error option to `test local` command (#2512)
This flags (which is `false` by default) tells the framework whether to
clean up (or not) if the test encountered errors.

With the current behaviour, the framework will always clean up the
resources it created. This is fine in most cases, however, in CI,
whenever there's an error, we might not get the logs of the workloads
and operator after the cleanup. So this option enables those resources
to linger until the logs have been acquired.

Co-Authored-By: Camila Macedo <cmacedo@redhat.com>