[helm-operator] Fix reconcilation of resources with jsonpatch
This PR updates the jsonpatch library to fix an issue during reconcilation of arrays.

The old version had an incomplete implementation for json patches of arrays.
A diff between
```
containers:
    - name: container1
    - name: container2
```
and
```
containers:
    - name: container1
```

created 3 patch operations:
- remove containers/0
- remove containers/1
- add containers/0

This causes an issue because we are ignoring all "remove" operations to allow manual modifications.

Helm tries to apply
```
containers:
    - name: container1
    - name: container2
    - name: container1
```

```
invalid: spec.template.spec.containers[0].name: Duplicate value:
"container1"
```

The new jsonpatch version does only return a `remove containers/1`
operation (which we are still ignoring because it could be a manual
change, but does not fail)