Improve Ansible User Guide memcached-operator example (#2774)
Improves the Ansible Operator SDK User Guide to improve usability.
This change to the documentation instructs the user to abstract an
environment variable, DOCKER_IMAGE, which stores the path to the
quay.io docker image.  Furthermore, the user is instructed to create
the DOCKER_IMAGE value based on his or her QUAY_USERNAME.  Although
this might seem implicit, I failed to recognize the importance of
using my own quay.io repository the first time, as I quickly
copied/pasted.

Lastly, the user guide was updated to instruct the user that the
memcached-operator repository should be made public.  Beginners,
such as myself, will benefit from this knowledge as without
changing the "Repository Visibility", the kubernetes deployment will
fail with an ImagePullError.

Signed-off-by: Ryan Goulding <rgouldin@redhat.com>