*: consolidate helm watch file management
This commit updates the `operator-sdk new` and operator-sdk add api`
commands to use the same function to create and append to Helm's
watches.yaml file.

This is a follow-up from a previous commit where we noticed that we
were using different mechanisms for updating the watches.yaml file.

  - `operator-sdk new` was using text/template to create the file.
  - `operator-sdk add api` was using yaml.Marshal to update the
    file.

Now, text/template is used for both cases, and removes the dependency
on `gopkg.in/yaml.v2` that was used for updating the helm's watches.yaml