Add support for cluster-scoped CRDs for kube-state-metrics (#2809)
Cluster-scoped CRD’s were not supported in the `pkg/kube-metrics` implementation since it was always looking for the GKV's in the List of namespaces informed which cause the issue "resource not found" since the CRD is not in any specific namespace. 

With the changes made here `pkg/kube-metrics` is able to List and Watch cluster-scoped CRDs.  Closes: #2707