pkg/helm/release/manager.go: refactor to use ThreeWayMergePatch for native k8s objects (#2869)
A follow-up PR to #2808 and #2777. @joelanford

In this PR we are using the Kubernetes strategic three-way merge patch for patches by the helm-operator for native Kubernetes objects.
This ensures that we are using the correct merge strategy that is defined in the OpenAPI schema of the object. #2808

However, we do not have the required schema information for custom resources(CRs). So we need to fall back to the previous manual json-patch.
To detect CRs we are using the same method as helm https://github.com/helm/helm/blob/b21b00978539ea8270e6b672bc1e6bc07af61475/pkg/kube/client.go#L391

Co-authored-by: Joe Lanford <joe.lanford@gmail.com>
Co-authored-by: Henning Surmeier <h.surmeier@mittwald.de>