CONTRIBUTING.md: fix broken hyperlink to the developer docs
This fixes the broken hyperlink to the developer docs in the
CONTRIBUTING.md file, which points to a former docs directory
inside the repo. Since that no longer exists, and the docs
have been moved onto the website, this fixes the hyperlink to
point to the same docs on the new website.

Fixes #2974