plugin: write Makefile recipes for `generate bundle` subcommand (#2860)
The Init plugin appends 'bundle' and 'bundle-build' recipes to the Makefile
to generate a bundle on-disk and build a bundle image, respectively, using
'generate bundle'. It also generates a kustomization.yaml in config/bundle/
so a user can pass output of 'kustomize build config/bundle' as generator
input. The CreateAPI plugin scaffolds config/samples/kustomization.yaml
so CR examles are included in bundle generator input