[olm-cmd] Package SDK with a default version of OLM
This PR uses go-bindata to embed a default version of OLM, when
user does not have any preference, so that SDK need not fetch
the manifests from github everytime.
The OLM manifests in this PR have the version "v0.15.1". With every
release SDK the olm-bindata-vX.X.X.go intends to be updated
with the most successful released version of OLM manifests.