upgrade lint version and fix (#3270)
Upgrade lint version used to the latest one
Fix go fmt check. See that we are just changing the fmt of 3 lines according to go rules which are valid for 1.13 and 1.14. See that Travis is using 1.13 and passed in the git diff test in the sanity check as well. (btw, done automatically when we run make lint-fix too).