feat: add kb scaffold utilities for plugins in internal/kubebuilder (#3305)
**Description**
- Just add the KB internal in  internal/kubebuilder
- change hack/check-license.sh to allow kb license

**Motivation**
SDK is in a process to be integrated with KB which means that its project layouts will be aligned. More info : [Integrating Kubebuilder and Operator SDK](https://github.com/kubernetes-sigs/kubebuilder/blob/master/designs/integrating-kubebuilder-and-osdk.md). As the first step, for we getting starting to working on in the Ansible/Helm plugins with the new layout we defined the strategy to just copy these helpers which will allow us to have the same features/utilities required to do the scaffolds.   

Note that these utilities are not exported by kubebuilder, in this way,  since Ansible/Helm will be also plugins that will do similar scaffolds and following a similar design they will be useful in the next steps. For a further understanding see the WIP: https://github.com/operator-framework/operator-sdk/pull/329 and the doc [Extensible CLI and Scaffolding Plugins](https://github.com/kubernetes-sigs/kubebuilder/blob/master/designs/extensible-cli-and-scaffolding-plugins-phase-1.md).