doc: add migration guide for helm new layout - Not visible in website  (#3342)
**Description of the change:**
- Add doc with the steps required that need is followed by the users migrate their project for the new layout. 
IMPORTANT: It is not visible on the website. The follow-up PR #3326 will make it visible. 

**Motivation for the change:**
SDK is in a process to be integrated with KB which means that its project layouts will be aligned. More info : [Integrating Kubebuilder and Operator SDK](https://github.com/kubernetes-sigs/kubebuilder/blob/master/designs/integrating-kubebuilder-and-osdk.md). 

As part of this process, we are creating the Helm plugin. See PR https://github.com/operator-framework/operator-sdk/pull/3295. Also, we are updating the steps in the doc in the PR https://github.com/operator-framework/operator-sdk/pull/3326 and here we are adding a new doc which will help users to migrate their projects.