feat: allows format the command text just for the CLI docs generation (#3426)
**Description:**
Add func to parse and replace the CLI command text for just for the docs generation in : `hack/generate/cli-doc/gen-cli-doc.go`. Note that it does not change the CLI texts outputs for the tool. 

Note that here it just replaces "<", for ""&lt;"" and ">" for "&gt;" to allow this value shows properly in the CLI docs synopsis only. 

**Motivation for the change:**
Closes: #3004