scorecard,generate-csv: consider api "group" while checking for known types (#3334)
Comparing only "kind" of the resource might lead to conflict where two
resources might have same kind but in different api groups.

For example an operator could provide a custom resource with name
"Deployment" in it's own group. On generating csv for this operator
might not consider the provided CR for filling 'alm-examples'.