Makefile,.travis.yml: Remove docker image builds for s390x architecture (#3710)
Travis CI builds for s390x are failing. If left in the build, we will
be unable to complete a release.

Therefore, until further notice, no s390x images will be built as part
of the release automation. If necessary, we can manually build and push
s390x images on a case-by-case basis.