Use rpm installed python-setuptools (#3830)
Using --ignore-installed with pip causes it to ignore the fact that
setuptools has already been installed by rpm. This results in the newest
version of setuptools (50.0.0) to be pulled in.

Setuptools 50 includes a backwards incompatible change to the way that
non-binary installs happen, which breaks the installation of
ruamel.yaml.clib on architectures that must install from source
because there is not a built wheel.