[olm] Store previous versions of olm manifests
This PR introduces an internal optimization of not going to github
and downloading the olm manifests when 'operator-sdk olm install'
command is used. Previous version of olm manifest is stored as
bindata and packaged along with sdk binary. When user specifies
any of this version, the manifests are fetched locally and
applied to the cluster.