doc: add base doc that clarifies how users can test their projects (#3823)
**Description of the change:**
doc: add base doc to clarifies how users can test their projects

**Motivation for the change:**
Many users have raised these questions. The goal of this doc is to provide the basic information and the links/references for they are able to move forward. It can be improved by the community and/or in the post 1.0 doc plans. 

Closes: https://github.com/operator-framework/operator-sdk/issues/3511