generate: add `--package` and permit extra-project use of `<bundle|packagemanifests>` (#4074)
This commit allows CSVs to be passed into 'generate <bundle|packagemanifests>'
commands and updated verbatim, and removes their dependency on
a project config. Version is used to overwrite CSV name and
`.spec.version` if `--version` is set, otherwise all metadata
fields are unmodified. This allows, for example, `.spec.replaces`
to be set or unset depending on what is in the input CSV.
The `--package` flag was added to 'generate' subcommands to
specify a package name different than the project config's
`projectName`, which is used by default when available.

internal/cmd/operator-sdk/generate: remove WithBase() option and read
base directly from file if one was not passed into stdin. Also add
the `--package` flag, which is required if not running in a project

internal/generate/clusterserviceversion: remove dependency on
bases in generator, change OperatorName -> PackageName

internal/generate/collector: add CSV type to collector

Co-authored-by: Eric Stroczynski <ericstroczynski@gmail.com>