Ignore CVEs 2021-3533 and 2021-3532 in Ansible 2.9 base image (#5465)
* Ignore CVEs 2021-3533 and 2021-3532 in base image

The following CVEs are causing `pipenv check` step of image build to
fail:

https://nvd.nist.gov/vuln/detail/CVE-2021-3533
https://nvd.nist.gov/vuln/detail/CVE-2021-3532

Both of these CVEs are against Ansible 2.9. The Ansible team is aware of
them, but they require deliberately insecure configuration from
the user and the Ansible team considers them invalid CVEs. There are no
plans for these to be fixed in the 2.9.z stream and will not cause a
security problem with Ansible-based operators.

Signed-off-by: austin <austin@redhat.com>

* Rerun sanity tests in 2022

Signed-off-by: austin <austin@redhat.com>