Return a bool instead of a hard exit.
The PrintWithFormat method would hard exit if there was an error in the
test. Instead of returning a special error for this case, we added a
boolean to indicate if there is a failed result. This allows the cmd to
handle this case and we no longer have a hard exit in the middle of a
library function.

Signed-off-by: jesus m. rodriguez <jmrodri@gmail.com>