Document setting logging with Kustomize
The Zap logger framework does not appear to support using environment-variables for setting the log level etc, so users need to be able to modify the command line to set log levels.

Kustomize is very commonly used to deploy operators. Document how to set the log level with a Kustomize patch, without having to copy the whole Deployment and hack it by hand.

Ideally the controller would support log level configuration in the environment and document it, so a simple `ConfigMap` with `envFrom` could be used to control logging instead. But in the absence of this, documenting how to set up log levels somehow is desirable.

Signed-off-by: Craig Ringer <craig.ringer@enterprisedb.com>