Fix scorecard XUnit result schema (#5774)
Error and failure cases are currently reported incorrectly. Tools that
consume the XUnit schema expect the result element to be named with its
type and also have an attribute with its type. This commit changes the
output for the result from the incorrect format:
`<result type="failure">..></result>` to the expected format:
`<failure type="failure">...</failure>`.

Signed-off-by: Ryan King <ryking@redhat.com>