OCPBUGS-1666: Expose flag to enable/disable PodSecurity (#6062)
* OCPBUGS-1666: Expose flag to enable/disable PodSecurity

Added --security-context-config flag to enable seccompprofile.
It defaults to enabled to support k8s 1.25. You can disable it
with --security-context-config=legacy

* Removed the "default: restricted" from the help text.

This was being duplicated because we had it in the text but were not
setting the value to a default value. Once we set the value to the
default cobra realized this and would output "(default: restricted)".

So removing the manually entered text fixes the duplicate.

Signed-off-by: jesus m. rodriguez <jesusr@redhat.com>

* fix test-sanity: update docs & fix test.
* Ignoring error from Set call in test
* Update .cncf maintainers
* Update run bundle(-upgrade) CLI docs
* Add allowed values to the security-context-config flag.
* Default to Restricted, and add the default value in the help text.
* Update changelog to indicate that we're updating run bundle subcommand
* Fix typo in changelog

Signed-off-by: jesus m. rodriguez <jmrodri@gmail.com>