re-orders deploy steps and clarifies instructions (#6109)
Previously, if you ran the suggested command `kustomize build
config/default | kubectl apply -f -`, that would render an incorrect
manifest. That is because when the user changed the value of
`IMAGE_TAG_BASE` and `IMG` in the Makefile during a previous step, that
change has not yet been applied to kustomize manifests at the time of
running the suggested command. So the wrong image reference ends up
being rendered, and the manifests get applies to the cluster. `make
deploy` must be run first, which updates the kustomize manifest.

This change clarifies that running `kustomize build config/default` is
part of explaining to the reader how things work, is not a required step
that should apply anything to the cluster, AND importantly should not
be run until after `make deploy`.