fixes broken "make run" for ansible operators (#6110)
* fixes broken "make run" for ansible operators

Previously, an ansible operator would be scaffolded with a Makefile
whose "run" target had incorrect syntax. It would run this command:

`ANSIBLE_ROLES_PATH="$(ANSIBLE_ROLES_PATH):$(shell pwd)/roles" $(ANSIBLE_OPERATOR) run`

That command failed to set the `ANSIBLE_ROLES_PATH` variable correctly
because it was using Makefile syntax in a target's recipe, which gets
interpreted as shell script.

This commit moves the variable setting statement, which uses Makefile
syntax, out of the recipe. It also uses the standard `?=` operator to
only set the variable if it is unset.

Signed-off-by: Michael Hrivnak <mhrivnak@hrivnak.org>

* Fix test-sanity

Signed-off-by: jesus m. rodriguez <jesusr@redhat.com>

Signed-off-by: Michael Hrivnak <mhrivnak@hrivnak.org>
Signed-off-by: jesus m. rodriguez <jesusr@redhat.com>
Co-authored-by: jesus m. rodriguez <jesusr@redhat.com>