Set REST timeout when timeout cli arg is provided
The timeout arg sets the overall timeout for the full command to finish,
such as `run bundle`, but it doesn't affect the timeout of individual
API server requests. This remains the default timeout specified
in the API server configuration. This means that even if a user wants
`run bundle` to timeout after 5m, the command may return earlier due to
an individual request's timeout being hit.

This commit fixes that behavior by setting the Timeout value in the
client cfg to the value passed in by the user, this will prevent the
individual request from timing out before the desired timeout.

Signed-off-by: Sebastian Soto <ssoto@redhat.com>